+++ CHANGELOG.md
@@ -12,6 +12,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [SQL] Added support for SQL functions `DATE_TRUNC`, `TIME_TRUNC`, and
   `TIMETSTAMP_TRUNC`. ([#2227](https://github.com/feldera/feldera/pull/2227))
 - [WebConsole] separate SQL and Rust compilation status into separate indicators
+- [WebConsole] enable creating pipeline with from a preset demo or a custom program via link
+- [WebConsole] redirect to a desired page after login
 
 ## [0.23.0] - 2024-08-06